requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==3.0.0
Django==4.2.22
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0